pyproject.toml
[project]
name = "multi-server-mcp-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=3.1.0",
    "langchain-groq>=1.1.2",
    "mcp-use>=1.6.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
